Added working with double values.

+++ DManager/Models/DebtModel.cs
@@ -3,7 +3,7 @@
     public class DebtModel
     {
         public string Name { get; set; }
-        public int DebtChange { get; set; }
+        public double DebtChange { get; set; }
         public string Description { get; set; }
     }
 }
